refactor: move globalStyles, set default Text type
