+++ applications/launchpad_v2/src/components/Text/styles.ts
@@ -1,0 +1,3 @@
+import styled from 'styled-components'
+
+export const StyledText = styled.p``
