+++ applications/launchpad_v2/src/components/Text/types.ts
@@ -2,13 +2,13 @@ import { ReactNode } from 'react'
 
 /**
  * @typedef TextProps
- * @prop {'header' | 'subheader' | 'defaultHeavy' | 'defaultMedium' | 'defaultUnder' | 'smallHeavy' | 'smallMedium' | 'smallUnder' | 'microHeavy' | 'microRegular'  | 'microOblique' } type - text styles
+ * @prop {'header' | 'subheader' | 'defaultHeavy' | 'defaultMedium' | 'defaultUnder' | 'smallHeavy' | 'smallMedium' | 'smallUnder' | 'microHeavy' | 'microRegular'  | 'microOblique' } [type] - text styles
  * @prop {ReactNode} children - text content to display
  * @prop {string} [color] - font color
  */
 
 export interface TextProps {
-  type:
+  type?:
     | 'header'
     | 'subheader'
     | 'defaultHeavy'
