+++ applications/launchpad_v2/src/styles/styles/index.ts
@@ -1,6 +1,6 @@
 import colors from './colors'
 import gradients from './gradients'
-import typography from '../typography'
+import  typography from './typography'
 
 const styles = {
   colors,
