+++ applications/launchpad_v2/src/styles/styles/typography.ts
